+++ packaging/homebrew/kapsis.rb
@@ -11,9 +11,9 @@ class Kapsis < Formula
 
   # Stable release - automatically updated by CI on each release
   # RELEASE_VERSION_MARKER_START - Do not remove, used by CI
-  url "https://github.com/aviadshiber/kapsis/archive/refs/tags/v1.4.0.tar.gz"
-  sha256 "28ed78d34748384aae12fa006a032d1834962764603906f45e394ee56648adf3"
-  version "1.4.0"
+  url "https://github.com/aviadshiber/kapsis/archive/refs/tags/v1.5.1.tar.gz"
+  sha256 "66418fb6a0b37e4f95fe30494205848d27609da08bfb6c49bdf9b3bff1c7ffb1"
+  version "1.5.1"
   # RELEASE_VERSION_MARKER_END
 
   # Homebrew livecheck - detects new releases automatically